Pin clang-format to v11 for now
The last nightly of clang-format (20201210052626+8a874a42777) changed
how C++ embedded in raw strings are formatted (for the better). This
introduces an annoyance where editors and tools that integrate with
libFormat might not have caught up (as is the case of CLion). Given that
we make pretty extensive use of raw strings in this project, and that
LLVM 12 is still in flux, pin down clang-format to v11 for now.