Move idempotence check in base case to Annotate*() methods
This somewhat simplifies the matchers in the base case, but more
importantly it sets a clearer pattern for the base case that what we're
looking for shouldn't depend on the presence of any annotation.

A bit more subtly, this also ensures correctness for cases where we
annotate more than what we find (e.g. when functions are involved, we
will look for all previous declarations, and try to find overridden
virtual methods).

For performance reasons, we have to keep the annotation check in the
matcher for the two base case rules for fields.