Extract common logic of making sets from matchers
This eliminates all ceremonies necessary to construct a VarSet or a
FieldSet. As part of this effort, we've also extracted the logic of
converting the sequence of BoundNodes to that of AST node's (e.g.
VarDecl). This should make the common case of using matchers with a
single binding a lot more fluent.