Ensure param type consistency among redecls
So far we've been inferring function return types and parameter types by
inspecting the body of functions. One subtle assumption this allows is
that we can redeclarations of the function we annotate will practically
be *all* in the same translation unit (you change the definition, and
there's likely one previous declaration from the header, then we factor
in virtual overrides).

This assumption will soon not hold as we introduce rules that infer
parameter types from the call site. There it's often the case that the
only declaration seen at the call site translation unit is one from the
header, leaving the definition unannotated.

In fact, this assumption has been silently broken when we propagate
virtual function return types: one TU that propagates a virtual
function's return type annotation "up"  will often only be able to
annotate the overridden method's declaration in a header.

We fix this by sweeping through all functions with annotated parameters,
and spread the annotations among redecls.