Propagate virtual function return types to redecls
This commit simplifies somewhat the logic of propagating virtual
function return type annotations. And it fixes a subtle bug: function
return types are merged by the parser, which means our original strict
matcher (find pairs where one is annotated and the other is not) will
miss cases where the header is annotated (by propagation in another TU)
but the main definition is not annotated: in those cases it's likely
that our old strict matcher, e.g. "not annotated" will miss the
definition because of the merged return type.

Simplify the match by only looking at annotated return types, and spread
it up and down the virtual overrides, regardless of the annotation
status of the "other" declarations.