Expand moved owner detection to descendant calls in returned expression
This expands on the idea of the previous commit , but with a new
insight: every variable referenced in the full expression that's being
returned is moved here, not just the top level expression. That means
every argument passed to an owner parameter in this full expression is a
moved owner.