Avoid using unless(equalsBoundNode) with forEach* matchers
The highly efficient implementation of equalsBoundNode makes its
behavior intricate and confusing in the presence of a prior, multi-match
binding (e.g. bindings inside forEach* traversal matchers, as opposed to
most of the bind-first-match kind of traversal): a node matches
equalsBoundNode as long as *any* of the prior match contains the id-node
binding. That brings us to the counter-intuitive behavior of
unless(equalsBoundNode()): it's tempting to interpret that matcher as
"does not equal *one of* the previous match", but it actually matches
nodes that equal *none* of the prior matched bindings.

I started replacing the part of the matcher that detects a second
reference to a variable in a return statement. For domain-specific
reasons, the existing matcher *actually* works, but continuing to use
this pattern will only confuse our future selves. Switch to a
hand-rolled inner loop match.