Replace HasBoundStmt* with StmtIs* matchers. NFC.
The new duo of matchers directly address the original need of matching a
pattern of addref-return (or init-then-addref), and it removes the
awkward need to bind a statement just so that we can use the matcher.

An upcoming change will also find the new StmtIsImmediateBefore matcher
useful, to be defensive against mistaking an owner for a poiner.