Be judicious with annotating vars based on instantiation information
We experimented with the simpler sledgehammer of just excluding vars in
template instantiation in a single place in AnnotateOneVar() but that
didn't work out very well: it missed a lot of useful annotations
produced with sufficient context (e.g. initializing a var with new() in
a member function of a class template).

Instead, we take pains in inspecting each place we potentially annotate
vars and determine whether we have enough contextual knowledge. In
places where the context is scarce, defend against vars in template
instantiations.