Expand pointer detection to variables with non-pointer initializers
We can drop the initializer check because we're careful to avoid
variables that are "plausibly owners", including variables that are
passed as non-pointer function arguments. That means whatever is left
cannot be an owner, or it will leak. (In a prior commit we introduced a
"leaked" annotation to facilitate manual marking of such variables, most
likely in tests.