Simplify "single decl" handling
Centralize the escape from multi-decl into AnnotateOneVar() and remove
variants of "SingleDecl()" sprinkled throughout the code. Now we never
need to explicitly match variables that are declared alone, and instead
when we escape when the annotation function encounters such variables.
This lays the ground work for an upcoming change where we add a pass to
annotate declarations that contain multiple variables.