Match current node in tail call analysis
We replace hasDescendant with SelfOrHasDescendant, forEachDescendant
with findAll. This is purely a defensive move: currently both of those
matchers are matching the descendants of ReturnStmt, which won't match
any expressions. However, this paves the way for the upcoming change
where we also recognize the last statement in a void function as a tail
call (we likely will never get to the CFG nirvana, but this should get
us a quick win).