Find more "last statement" using CFG
We started out with just return statements, then bolted on (absolute)
"last expression" for void-returning functions. We finally bit the
bullet of using CFG and uniformly handle all "last statements" from a
control-flow perspective: a "last statement" is the last statement in a
basic block that is immediately (and unconditionally) before the
function exit.

To keep the scope of this change manageable, we still bail on
constructors: parameters used in initializers for owner fields are still
incorrectly reported as pointers if they're referenced in the last
expression in the constructor body (often in an assertion).