Speed up propagation by feedback
As we go along annotating the source code, remember some of the
inferences we've made (which vars are owners, which are pointers), and
feed that back into subsequence decisions. This means we can loop over
the same process until no further progress is made. This makes each run
(executing our binary over a single source file) slower, but it reduces
the number of iterations from 15 to just 6.

Observation: the vast majority of time of execution is spent *before*
out code even gets run: it's expensive to parse the source and generate
an AST, so let's squeeze more value out of it.