Expand the last-use propagation to many more places
We now not only analyze the literal "last" statement before exiting a
function, but *every* statement that we're certain will lead to the exit
(without branching), while associating with each statement a set of
variables that are last used in it. We also explicitly analyze
constructor initializers for last-use.

This expands on the work done in d6e9a6a3b3a42c33 (which is an extension
of earlier commit d15bef847b982f47 that introduced the powerful tool of
control-flow graph).

Note the new code does *not* subsumes the feature (and eliminates code)
added in commit 77215ec670991ea4, where we can infer parameter ownership
based on the field they are initializing in a constructor initializer.
It might be counter-intuitive at first, but parameters initializing
fields sometimes not last-used in the constructor initializer, but
instead pointer-used in the body of the constructor (e.g. an assert).