[convert] remove unused AutoRef objects
We have a particular pattern of using AutoRef like the following

CAutoRef<T> at;
owner<T*> ot = ...;
at = ot;

And then "at" just chills until the end of scope. This will be
extraneous once we convert "ot" to a Ref<T>. Here find this pattern and
erase the variable "at" and the assignment.