Generate release / debug versions of the compilation database
Due to limitations of my CFG-based approach to detecting "moves", an
assertion can stop us from detection of any "last use" of a variable
before the assert. To work around this, run the annotation code once in
release and once in debug.

This might add a little overhead to the whole process, and if it becomes
apparently unbearable we still have room for improvement: we can
generate the two compilation databases just once, stash them aside, and
subsequently only rerun configure (to regenerate headers), and copy back
the recorded version of the compilation database (that should save time
on building the translator).