Use VPATH for gpopt
This commit gets rid of the kludge of rebuilding and regenerating the
compilation database for the main GPDB, and instead uses two
pre-generated VPATH, each containing complete generated headers and a
compile_commands.json database. This brings the logic of processing the
gpopt translator a lot closer to that of gporca.